+++ src/DefaultBuilder/src/StaticWebAssetsWebHostBuilderExtensions.cs
@@ -1,7 +1,6 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System.IO;
 using Microsoft.AspNetCore.Hosting;
 
 namespace Microsoft.AspNetCore
@@ -21,7 +20,7 @@ namespace Microsoft.AspNetCore
         {
             builder.ConfigureAppConfiguration((context, configBuilder) =>
             {
-                StaticWebAssetsHelper.UseStaticWebAssets(context.HostingEnvironment);
+                StaticWebAssetsLoader.UseStaticWebAssets(context.HostingEnvironment);
             });
 
             return builder;
